Windows doesn't let a socket outlive the process
So, just disconnect clients and let them race to be the new server.

It's not quite as performant, but it actually works, and is still better
than a polling-based fs O_EXCL locking mechanism.

Tests still don't even begin to pass on Windows, but a lot of the cases
don't matter as much anyway, because they're actually impossible with
windows pipes.